Moving required scopes into auth file
